pyproject.toml
[project]
name = "mcp-raven"
version = "0.1.0"
description = "MCP memory server — LLM agents that remember and recall for you"
requires-python = ">=3.11"
dependencies = [
    "mcp-agent[openai]>=0.2.6",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
